tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-duration_start_end_events.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-duration: 0s, start and end events</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-duration-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="Animation events are fired even when
   animation-duration is 0s">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes grow {
         from {
            width: 100px;
         }
         to {
            width: 200px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         position: relative;
      }
      .test {
         animation-name: grow;
         animation-duration: 0s;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var startPassed = false;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 500);
         div.addEventListener('animationstart', function() {
            startPassed = true;
         });
         div.addEventListener('animationend', function() {
            clearTimeout(timeout);
            test(function(){assert_true(startPassed);}, testName); done();
            log.data = (startPassed ? 'PASS' : 'FAIL');
         });
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue block below should stay square except possibly for a brief flash,
   and the next paragraph should say "PASS".</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
